// src/types/question.enum.ts
/**
 * Enumeration representing mathematical operation symbols and a blank placeholder.
 * @enum {string}
 */
export enum MathSymbol {
  Addition = '+',
  Subtraction = '-',
  Multiplication = '*',
  Division = '/',
  Equals = '=',
  Blank = '?',
}

/**
 * Enum representing difficulty levels for the game.
 * @enum {string}
 */
export enum Difficulty {
  EASY = 'easy',
  MEDIUM = 'medium',
  HARD = 'hard',
}